servidor/requirements.txt
flask
python-dotenv
stripe
flask-cors
